Fix signed vs unsigned kuid comparisons
Kuids may be represented in signed or unsigned form by differente Trainz versions (TS12 vs TRS19).

Also, the same kuid might be represented as a signed integer in binary form, but as an unsigned integer in string form.

Make sure we compare the actual kuid binary form, not the string representation.

For example <kuid:224567:2943774475> and <kuid:224567:-1351192821> represent the same kuid.